Add ConcreteComponent type
