+++ ViewComponents/Component.swift
@@ -8,54 +8,7 @@
 
 import UIKit
 
-protocol ComponentType {
-    var styles: Set<AnyStyle> { get }
-    func configure(view: UIView)
-    func isEqual(to other: ComponentType) -> Bool
-    func diffChanges(from other: ComponentType) -> ComponentType
-}
-
-struct ChildComponent: ComponentType, Equatable {
-    let component: ComponentType
-    let access: (UIView) -> UIView
-    
-    init<V: UIView, T: UIView>(component: Component<V>, _ access: @escaping (T) -> V) {
-        self.init(component, { view in access(view as! T) })
-    }
-    
-    init(_ component: ComponentType, _ access: @escaping (UIView) -> UIView) {
-        self.component = component
-        self.access = access
-    }
-    
-    var styles: Set<AnyStyle> {
-        return component.styles
-    }
-    
-    func configure(view: UIView) {
-        component.configure(view: access(view))
-    }
-    
-    func isEqual(to other: ComponentType) -> Bool {
-        guard let other = other as? ChildComponent else { return false }
-        return component.isEqual(to: other.component)
-    }
-    
-    static func == (lhs: ChildComponent, rhs: ChildComponent) -> Bool {
-        return lhs.component.isEqual(to: rhs.component)
-    }
-    
-    func diffChanges(from other: ComponentType) -> ComponentType {
-        guard let otherComp = other as? ChildComponent else { return other }
-        return diffChanges(from: otherComp)
-    }
-    
-    func diffChanges(from other: ChildComponent) -> ChildComponent {
-        return ChildComponent(component.diffChanges(from: other.component), other.access)
-    }
-}
-
-public struct Component<T: UIView>: ComponentType {
+public struct Component<T: UIView>: ConcreteComponentType {
     public let styles: Set<AnyStyle>
     let children: [ChildComponent]
     
@@ -87,11 +40,6 @@ public struct Component<T: UIView>: ComponentType {
 }
 
 extension Component {
-    func diffChanges(from other: ComponentType) -> ComponentType {
-        guard let otherComp = other as? Component<T> else { return other }
-        return diffChanges(from: otherComp)
-    }
-    
     public func diffChanges(from other: Component<T>) -> Component<T> {
         let newStyles = other.styles.subtracting(styles)
         let newChildren = other.children.lazy.enumerated().map({ current -> ChildComponent in
@@ -111,7 +59,7 @@ extension Component: ComponentConvertible {
     }
 }
 
-extension Component: Equatable {
+extension Component {
     func isEqual(to other: ComponentType) -> Bool {
         guard let other = other as? Component<T> else { return false }
         return self == other
